+++ kernel.go
@@ -165,7 +165,7 @@ const (
 // Raw strings are with a terminating NUL character. For convenience, use KernelInfoString().
 //
 // See also: https://registry.khronos.org/OpenCL/sdk/1.2/docs/man/xhtml/clGetKernelInfo.html
-func KernelInfo(kernel Kernel, paramName KernelInfoName, paramSize uint, paramValue unsafe.Pointer) (uint, error) {
+func KernelInfo(kernel Kernel, paramName KernelInfoName, paramSize uintptr, paramValue unsafe.Pointer) (uintptr, error) {
 	sizeReturn := C.size_t(0)
 	status := C.clGetKernelInfo(
 		kernel.handle(),
@@ -176,7 +176,7 @@ func KernelInfo(kernel Kernel, paramName KernelInfoName, paramSize uint, paramVa
 	if status != C.CL_SUCCESS {
 		return 0, StatusError(status)
 	}
-	return uint(sizeReturn), nil
+	return uintptr(sizeReturn), nil
 }
 
 // KernelInfoString is a convenience method for KernelInfo() to query information values that are
@@ -185,7 +185,7 @@ func KernelInfo(kernel Kernel, paramName KernelInfoName, paramSize uint, paramVa
 // This function does not verify the queried information is indeed of type string. It assumes the information is
 // a NUL terminated raw string and will extract the bytes as characters before that.
 func KernelInfoString(kernel Kernel, paramName KernelInfoName) (string, error) {
-	return queryString(func(paramSize uint, paramValue unsafe.Pointer) (uint, error) {
+	return queryString(func(paramSize uintptr, paramValue unsafe.Pointer) (uintptr, error) {
 		return KernelInfo(kernel, paramName, paramSize, paramValue)
 	})
 }
@@ -258,7 +258,7 @@ const (
 // Raw strings are with a terminating NUL character.
 //
 // See also: https://registry.khronos.org/OpenCL/sdk/1.2/docs/man/xhtml/clGetKernelWorkGroupInfo.html
-func KernelWorkGroupInfo(kernel Kernel, device DeviceID, paramName KernelWorkGroupInfoName, paramSize uint, paramValue unsafe.Pointer) (uint, error) {
+func KernelWorkGroupInfo(kernel Kernel, device DeviceID, paramName KernelWorkGroupInfoName, paramSize uintptr, paramValue unsafe.Pointer) (uintptr, error) {
 	sizeReturn := C.size_t(0)
 	status := C.clGetKernelWorkGroupInfo(
 		kernel.handle(),
@@ -270,7 +270,7 @@ func KernelWorkGroupInfo(kernel Kernel, device DeviceID, paramName KernelWorkGro
 	if status != C.CL_SUCCESS {
 		return 0, StatusError(status)
 	}
-	return uint(sizeReturn), nil
+	return uintptr(sizeReturn), nil
 }
 
 // KernelArgInfoName identifies properties of a kernel argument, which can be queried with KernelArgInfo().
@@ -352,7 +352,7 @@ const (
 // Raw strings are with a terminating NUL character.For convenience, use KernelArgInfoString().
 //
 // See also: https://registry.khronos.org/OpenCL/sdk/1.2/docs/man/xhtml/clGetKernelArgInfo.html
-func KernelArgInfo(kernel Kernel, index uint32, paramName KernelArgInfoName, paramSize uint, paramValue unsafe.Pointer) (uint, error) {
+func KernelArgInfo(kernel Kernel, index uint32, paramName KernelArgInfoName, paramSize uintptr, paramValue unsafe.Pointer) (uintptr, error) {
 	sizeReturn := C.size_t(0)
 	status := C.clGetKernelArgInfo(
 		kernel.handle(),
@@ -364,7 +364,7 @@ func KernelArgInfo(kernel Kernel, index uint32, paramName KernelArgInfoName, par
 	if status != C.CL_SUCCESS {
 		return 0, StatusError(status)
 	}
-	return uint(sizeReturn), nil
+	return uintptr(sizeReturn), nil
 }
 
 // KernelArgInfoString is a convenience method for KernelArgInfo() to query information values that are
@@ -373,7 +373,7 @@ func KernelArgInfo(kernel Kernel, index uint32, paramName KernelArgInfoName, par
 // This function does not verify the queried information is indeed of type string. It assumes the information is
 // a NUL terminated raw string and will extract the bytes as characters before that.
 func KernelArgInfoString(kernel Kernel, index uint32, paramName KernelArgInfoName) (string, error) {
-	return queryString(func(paramSize uint, paramValue unsafe.Pointer) (uint, error) {
+	return queryString(func(paramSize uintptr, paramValue unsafe.Pointer) (uintptr, error) {
 		return KernelArgInfo(kernel, index, paramName, paramSize, paramValue)
 	})
 }
